Decoupling coverage information from KmerNode class
KmerGraph and KmerNode were coupled with coverage information. This means that if we wanted to add coverage information when mapping a sample in pandora map or compare, we would need to copy the whole KmerGraph and change the coverage of all of its KmerNodes. The graph itself does not change, only the coverages associated to the nodes. This consumes more memory and slows the runtime a little bit. To improve on this, we now decoupled the coverage information from KmerNode. KmerGraph just holds the graph, with no coverage information. A new class in kmergraph.h, KmerGraphWithCoverage, adds coverage information to the KmerGraph.