Falling back to multithreading reads instead of samples in compare
There is a dependency in the main for loop (compare_main.cpp:309) that builds the sample pangraphs and the pangraph . This dependency is due to the e_rate variable: the e_rate can be updated from the previous iteration, which is considered in the next iteration. This is the only thing that is blocking us from multithreading in samples, but hopefully multithreading in reads will be also fast.