Solving a memory leak in pandora compare
There was a circular dependency with std::shared_ptr: pangenome::Read had references to pangenome::Node and pangenome::Node back to pangenome::Read. These references were done as std::shared_ptr, so their reference counting were never reaching 0 and thus neither pangenome::Node nor pangenome::Read could be cleaned up (see https://www.learncpp.com/cpp-tutorial/15-7-circular-dependency-issues-with-stdshared_ptr-and-stdweak_ptr/). This seems to be a big issue in compare, since we were not being able to clean up the hits of a sample after done with that sample. We still have some memory leaks to solve, but this was the largest one.