Refactoring VCFRecord
Changing the way samples infos and genotyped samples infos are stored. Now it uses a class SampleInfo and SampleInfos defined in sampleinfo.h/cpp. For now these classes are pretty trivial, but they will be refactored to contain more and more methods, and also to use composition instead of inheritance. However, this refactoring will be slow, since there are things that have higher priority.