Each SampleInfo now has the knowledge of how many alleles it contains. SampleIndexToSampleInfoTemplate is now safer.
The number of alleles should always be consistent with the coverage information. If the number of alelles change, their coverage is resized accordingly. To set the coverage information, it now needs to be exactly the number of alleles in the respective SampleInfo. This avoids possible bugs due to complex bookkeeping of these information.

SampleIndexToSampleInfoTemplate is now safer since we are now using composition instead of inheritance in std::vector. We control precisely how it is used. Only VCFRecord can change it. Other classes can access it.